Refactored ANT code
