+++ rf/fec.h
@@ -45,7 +45,9 @@ void roller_manager_tasks(void);
 
 void ant_evt_fec (ant_evt_t * p_ant_evt);
 
-void ant_fec_evt_handler(ant_fec_profile_t * p_profile, ant_fec_evt_t event);
+void fec_profile_setup(void);
+
+void fec_profile_start(void);
 
 #ifdef __cplusplus
 }
